+++ nginx-1.18.0/html/review/js/common.js
@@ -193,16 +193,16 @@ util.getRememberedBlogCard = function(id) {
 }
 
 const defaultShopTypes = [
-  { id: 1, name: "美食", icon: "/types/ms.png", sort: 1 },
-  { id: 2, name: "KTV", icon: "/types/KTV.png", sort: 2 },
-  { id: 3, name: "丽人·美发", icon: "/types/lrmf.png", sort: 3 },
-  { id: 10, name: "美睫·美甲", icon: "/types/mjmj.png", sort: 4 },
-  { id: 5, name: "按摩·足疗", icon: "/types/amzl.png", sort: 5 },
-  { id: 6, name: "美容SPA", icon: "/types/spa.png", sort: 6 },
-  { id: 7, name: "亲子游乐", icon: "/types/qzyl.png", sort: 7 },
-  { id: 8, name: "酒吧", icon: "/types/jiuba.png", sort: 8 },
-  { id: 9, name: "轰趴馆", icon: "/types/hpg.png", sort: 9 },
-  { id: 4, name: "健身运动", icon: "/types/jsyd.png", sort: 10 }
+  { id: 1, name: "Food", icon: "/types/food.svg", sort: 1 },
+  { id: 2, name: "Cafe", icon: "/types/cafe.svg", sort: 2 },
+  { id: 3, name: "Beauty", icon: "/types/beauty.svg", sort: 3 },
+  { id: 10, name: "Nails", icon: "/types/nails.svg", sort: 4 },
+  { id: 5, name: "Massage", icon: "/types/massage.svg", sort: 5 },
+  { id: 6, name: "KTV", icon: "/types/ktv.svg", sort: 6 },
+  { id: 7, name: "Family", icon: "/types/family.svg", sort: 7 },
+  { id: 8, name: "Bar", icon: "/types/bar.svg", sort: 8 },
+  { id: 9, name: "Party", icon: "/types/party.svg", sort: 9 },
+  { id: 4, name: "Fitness", icon: "/types/fitness.svg", sort: 10 }
 ];
 
 util.resolveTypeIcon = function(icon) {